// backend/src/audit-findings/entities/sample-enums.ts
export enum SampleType {
  DETAIL = 'DETAIL',
  PROCESS_CONTROL = 'PROCESS_CONTROL',
}

export enum SamplingMethod {
  RANDOM = 'RANDOM',
  SYSTEMATIC = 'SYSTEMATIC',
  JUDGMENTAL = 'JUDGMENTAL',
  MUS = 'MUS',
  STRATIFIED = 'STRATIFIED',
  HAPHAZARD = 'HAPHAZARD',
  BLOCK = 'BLOCK',
}

export enum TestResult {
  NOT_TESTED = 'NOT_TESTED',
  PASS = 'PASS',
  FAIL = 'FAIL',
  EXCEPTION = 'EXCEPTION',
  N_A = 'N_A',
}